added the possibility to return the mapping between original cell and its clone when calling the cloning method.
removed SCXMLEditor class and put it into SCXMLGraphEditor.
updated build script to reflect the change in the main method.

added functionality to display all the included content at the proper place but without the possibility of changing it.
(modified accordingly the save/open methods to behave as expected when included content is present/requested)